move defImages to run context
