+++ packages/core/src/types/prompt_template.d.ts
@@ -1229,6 +1229,7 @@ interface RunPromptContext {
         generator: string | RunPromptGenerator,
         options?: RunPromptOptions
     ): Promise<RunPromptResult>
+    defImages(files: StringLike, options?: DefImagesOptions): void
     defTool(
         name: string,
         description: string,
@@ -1490,7 +1491,6 @@ interface ContainerHost extends ShellHost {
 interface PromptContext extends RunPromptContext {
     script(options: PromptArgs): void
     system(options: PromptSystemArgs): void
-    defImages(files: StringLike, options?: DefImagesOptions): void
     defFileMerge(fn: FileMergeHandler): void
     defOutputProcessor(fn: PromptOutputProcessorHandler): void
     fetchText(
